修复 (int)offset 截断问题：SendFilePayload/SendUdpFileDataAsync fileOffset 改为 long
之前与 port=0 同时修改导致测试失败被误归因，现已分离验证通过。

+++ TrFileTransfer/TransferClient.cs
@@ -279,7 +279,7 @@ namespace TrFileTransfer
                 Log(string.Format("Chunk sending: {0} offset={1} size={2}",
                     fileName, offset, Utils.FormatSize(chunkSize)));
 
-                await SendFilePayload(stream, _filePath, chunkSize, fileName, ct, (int)offset);
+                await SendFilePayload(stream, _filePath, chunkSize, fileName, ct, offset);
 
                 var completeHandler = OnTransferComplete;
                 if (completeHandler != null) completeHandler();
@@ -287,7 +287,7 @@ namespace TrFileTransfer
         }
 
         private async Task SendFilePayload(NetworkStream stream, string filePath, long fileSize,
-            string displayName, CancellationToken ct, int fileOffset = 0)
+            string displayName, CancellationToken ct, long fileOffset = 0)
         {
             var sw = System.Diagnostics.Stopwatch.StartNew();
             long bytesSent = 0;
